requirements.txt
numpy
pandas
scikit-learn
matplotlib
tqdm
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
pillow
# pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu